+++ src/apps/content-editor/src/app/ContentEditor.js
@@ -27,7 +27,7 @@ import { CSVImport } from "./views/CSVImport";
 import "@zesty-io/core/vendor.css";
 
 import styles from "./ContentEditor.less";
-import AnalyticsDashboard from "./views/AnalyticsDashboard";
+import Analytics from "./views/Analytics";
 
 // Makes sure that other apps using legacy theme does not get affected with the palette
 let customTheme = createTheme(legacyTheme, {
@@ -154,11 +154,7 @@ export default function ContentEditor() {
               >
                 <div className={styles.ContentWrap}>
                   <Switch>
-                    <Route
-                      exact
-                      path="/content"
-                      component={AnalyticsDashboard}
-                    />
+                    <Route exact path="/content" component={Analytics} />
                     <Route
                       exact
                       path="/content/link/new"
