+++ src/apps/content-editor/src/app/views/Analytics/utils.ts
