+++ src/apps/content-editor/src/app/views/Analytics/views/AnalyticsDashboard/ByDayLineChart.tsx
@@ -18,7 +18,7 @@ import lineChartSkeleton from "../../../../../../../../../public/images/lineChar
 import {
   calculatePercentageDifference,
   findValuesForDimensions,
-} from "./utils";
+} from "../../utils";
 
 type Params = {
   modelZUID: string;
